Crate volta_fail

This crate provides a protocol for Volta's error handling, including a subtrait of the failure crate's Fail trait to manage the interface between errors and process exit codes.

The VoltaFail trait

The main interface for Volta errors is VoltaFail, which extends the Fail trait from the failure library with an additional method. The VoltaFail::exit_code() method allows each error type to indicate what the process exit code should be if the error is the reason for exiting Volta.

The VoltaError type and Fallible functions

The main error type provided by this crate is VoltaError. This acts more or less as the "root" error type for Volta; all Volta error types can be coerced into this type.

If you don't have any need for more specific static information about the errors that can be produced by a function, you should define its signature to return Result<T, VoltaError> (where T is whatever type you want for successful results of the function).

This is so common that you can use Fallible<T> as a shorthand.

Example

As a running example, we'll build a little parser for hex-encoded RGB triples. The type could be defined as a struct of three bytes:

#[derive(Debug)]
struct Rgb { r: u8, g: u8, b: u8 }

A function that decodes a single two-digit component could then use Fallible for its signature:

use volta_fail::Fallible;

// same as: fn parse_component(src: &str, i: usize) -> Result<u8, VoltaError>
fn parse_component(src: &str, i: usize) -> Fallible<u8> {
    // ...
}

Creating custom error types

To create an error type in Volta, add a #[derive] attribute to derive the Fail trait before the type declaration, and add a #[fail(display = "...")] attribute to construct the error message string.

Continuing with the running example, we could create an error type for running past the end of the input string:

Example

// required for `#[derive(Fail)]` and `#[fail(...)]` attributes
use failure::Fail;

use volta_fail::{ExitCode, VoltaFail};
use volta_fail_derive::*;

#[derive(Debug, Fail, VoltaFail)]
#[fail(display = "unexpected end of string")]
#[volta_fail(code = "InvalidArguments")]
struct UnexpectedEndOfString;

Throwing errors

The throw! macro is a convenient syntax for an early exit with an error. It can be used inside any function with a Result return type (often a Fallible<T>). The argument expression can evaluate to any type that implements a coercion to the declared error type.

Example

use volta_fail::throw;

fn parse_component(src: &str, i: usize) -> Fallible<u8> {
    if i + 2 > src.len() {
        // UnexpectedEndOfString implements VoltaFail, so it coerces to VoltaError
        throw!(UnexpectedEndOfString);
    }

    // ...
}

Using third-party error types

When using a third-party library that has error types of its own, those error types need to be converted to Volta errors. Since third party libraries have not been designed with Volta's end-user error messages in mind, third-party error types are not automatically converted into Volta errors.

Instead, this crate provides a couple of extension traits that you can import to add an with_context() method to errors (FailExt) or Results (ResultExt). This method will convert any third-party error to a Volta error.

Cause chains

Since errors get propagated up from lower abstraction layers to higher ones, the higher layers of abstraction often need to add contextual information to the error messages, producing higher quality messages.

For example, the ParseIntError produced by u8::from_str_radix does not tell the end user that we were parsing an integer in the context of parsing an RGB value.

To add contextual information to a lower layer's error, we use the with_context method and pass it a closure that takes a reference to the lower layer's error and uses it to construct a new higher-level error.

A powerful feature of with_context is that it saves the lower-level error message as part of a cause chain, which Volta's top-level can then use to produce in-depth diagnostics in a log file or for --verbose error reporting. Most error handling logic should not need to work with cause chains, so this is all handled automatically.

Example

use std::fmt::Display;
// add `with_context()` extension method to Results
use volta_fail::ResultExt;

#[derive(Debug, Fail, VoltaFail)]
#[fail(display = "invalid RGB string: {}", details)]
#[volta_fail(code = "InvalidArguments")]
struct InvalidRgbString { details: String }

impl InvalidRgbString {
    fn new<D: Display>(details: &D) -> InvalidRgbString {
        InvalidRgbString { details: format!("{}", details) }
    }
}

fn parse_component(src: &str, i: usize) -> Fallible<u8> {
    if i + 2 > src.len() {
        // UnexpectedEndOfString implements VoltaFail, so it coerces to VoltaError
        throw!(UnexpectedEndOfString);
    }

    // convert the std::num::ParseIntError into a VoltaError
    u8::from_str_radix(&src[i..i + 2], 16).with_context(InvalidRgbString::new)
}

Notice that you can use with_context to wrap any kind of error, including errors that may already be user-friendly. So you can always use this to add even more clarity to any errors. For instance, in our running example of an RGB parser, a higher layer may want to add context about which RGB string was being parsed and where it came from (say, the filename and line number).

Macros

throw
A temporary polyfill for throw! until the new failure library includes it.

Structs

VoltaError
The VoltaError type, which can contain any Volta failure.

Enums

ExitCode
Exit codes supported by the VoltaFail trait.

Constants

_IMPL_SERIALIZE_FOR_ExitCode

Traits

FailExt
An extension trait allowing any failure, including failures from external libraries, to be converted to a Volta error. This marks the error as an unknown error, i.e. a non-user-friendly error.

ResultExt
An extension trait for Result values, allowing conversion of third-party errors or other lower-layer errors into Volta errors.

VoltaFail
The failure trait for all Volta errors.

Type Definitions

Fallible
A convenient shorthand for Result types that produce VoltaErrors.